Long awaited update to many things (read)
Slight change to the way inputs are given to the neural network : rather than having different pieces represented by different values of the same number, we prefer to have well defined "virtual boards" for each piece type, black and white.
This leads to less confusion in the algorithm and, even if the number of input has significantly increased, we hope it will increase convergence rate.
Related : Changed piece numbering for it to be easier.
Also removed possible soft-lock inside tournaments if matches shut down unexpectedly.

Additionally, the range of ways to breed new generation has been increased to induce more diversity.

TODO : find a way to make the interface have arguments so we don't have 3 functions for nothing.